--- src/@types/basic.ts
@@ -1,5 +1,0 @@
-export type bytes = Uint8Array
-export type bytes32 = Uint8Array
-export type bytes16 = Uint8Array
-
-export type uint64 = number
