--- src/@types/libp2p.ts
@@ -1,10 +1,0 @@
-import type { bytes32 } from './basic.js'
-import type { NoiseExtensions } from '../proto/payload.js'
-import type { ConnectionEncrypter } from '@libp2p/interface'
-
-export interface KeyPair {
-  publicKey: bytes32
-  privateKey: bytes32
-}
-
-export interface INoiseConnection extends ConnectionEncrypter<NoiseExtensions> {}
